Relocate Attendance and Assignment class files
