+++ src/main/java/taa/assignment/StudentList.java
@@ -1,4 +1,4 @@
-package seedu.address.assignment;
+package taa.assignment;
 
 import java.util.ArrayList;
